Some refactoring of the area area index generator code. Changed default max magnification from 17 to 15.

+++ libosmscout-import/include/osmscout/import/GenAreaAreaIndex.h
@@ -64,7 +64,7 @@ namespace osmscout {
     bool WriteIndexLevel(const TypeConfigRef& typeConfig,
                          const ImportParameter& parameter,
                          FileWriter& writer,
-                         int level,
+                         size_t level,
                          std::map<Pixel,AreaLeaf>& leafs);
